+++ core/services/job/runner_integration_test.go
@@ -13,13 +13,12 @@ import (
 	"testing"
 	"time"
 
-	"github.com/smartcontractkit/libocr/commontypes"
-
 	evmconfigmocks "github.com/smartcontractkit/chainlink/core/chains/evm/config/mocks"
 	evmmocks "github.com/smartcontractkit/chainlink/core/chains/evm/mocks"
 	configtest2 "github.com/smartcontractkit/chainlink/core/internal/testutils/configtest/v2"
 	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/core/services/chainlink"
+	"github.com/smartcontractkit/chainlink/core/services/keystore/keys/ethkey"
 	ocr2mocks "github.com/smartcontractkit/chainlink/core/services/ocr2/mocks"
 	"github.com/smartcontractkit/chainlink/core/services/ocr2/validate"
 	"github.com/smartcontractkit/chainlink/core/services/srvctest"
@@ -34,7 +33,6 @@ import (
 	clhttptest "github.com/smartcontractkit/chainlink/core/internal/testutils/httptest"
 	"github.com/smartcontractkit/chainlink/core/internal/testutils/pgtest"
 	"github.com/smartcontractkit/chainlink/core/services/job"
-	"github.com/smartcontractkit/chainlink/core/services/keystore/keys/ethkey"
 	"github.com/smartcontractkit/chainlink/core/services/ocr"
 	"github.com/smartcontractkit/chainlink/core/services/ocrcommon"
 	"github.com/smartcontractkit/chainlink/core/services/pipeline"
@@ -56,11 +54,26 @@ import (
 var monitoringEndpoint = telemetry.MonitoringEndpointGenerator(&telemetry.NoopAgent{})
 
 func TestRunner(t *testing.T) {
-	config := cltest.NewTestGeneralConfig(t)
 	db := pgtest.NewSqlxDB(t)
-
-	keyStore := cltest.NewKeyStore(t, db, config)
+	keyStore := cltest.NewKeyStore(t, db, pgtest.NewQConfig(true))
+	kb, err := keyStore.OCR().Create()
+	require.NoError(t, err)
 	ethKeyStore := keyStore.Eth()
+	_, transmitterAddress := cltest.MustInsertRandomKey(t, ethKeyStore, 0)
+	require.NoError(t, keyStore.OCR().Add(cltest.DefaultOCRKey))
+
+	config := configtest2.NewGeneralConfig(t, func(c *chainlink.Config, s *chainlink.Secrets) {
+		t := true
+		c.P2P.V1.Enabled = &t
+		c.P2P.V1.DefaultBootstrapPeers = &[]string{
+			"/dns4/chain.link/tcp/1234/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju",
+			"/dns4/chain.link/tcp/1235/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju",
+		}
+		kbid := models.MustSha256HashFromHex(kb.ID())
+		c.OCR.KeyBundleID = &kbid
+		taddress := ethkey.EIP55AddressFromAddress(transmitterAddress)
+		c.OCR.TransmitterAddress = &taddress
+	})
 
 	ethClient := cltest.NewEthMocksWithDefaultChain(t)
 	ethClient.On("HeadByNumber", mock.Anything, (*big.Int)(nil)).Return(cltest.Head(10), nil)
@@ -75,9 +88,6 @@ func TestRunner(t *testing.T) {
 
 	require.NoError(t, runner.Start(testutils.Context(t)))
 	t.Cleanup(func() { assert.NoError(t, runner.Close()) })
-
-	_, transmitterAddress := cltest.MustInsertRandomKey(t, ethKeyStore, 0)
-	require.NoError(t, keyStore.OCR().Add(cltest.DefaultOCRKey))
 
 	t.Run("gets the election result winner", func(t *testing.T) {
 		var httpURL string
@@ -447,8 +457,6 @@ ds1 -> ds1_parse;
 		jb.MaxTaskDuration = models.Interval(cltest.MustParseDuration(t, "1s"))
 		err = jobORM.CreateJob(&jb)
 		require.NoError(t, err)
-		// Required to create job spawner delegate.
-		config.Overrides.P2PListenPort = null.IntFrom(2000)
 		sd := ocr.NewDelegate(
 			db,
 			jobORM,
@@ -481,8 +489,6 @@ ds1 -> ds1_parse;
 		jb.MaxTaskDuration = models.Interval(cltest.MustParseDuration(t, "1s"))
 		err = jobORM.CreateJob(&jb)
 		require.NoError(t, err)
-		// Required to create job spawner delegate.
-		config.Overrides.P2PListenPort = null.IntFrom(2000)
 
 		lggr := logger.TestLogger(t)
 		_, err = keyStore.P2P().Create()
@@ -506,8 +512,6 @@ ds1 -> ds1_parse;
 	})
 
 	t.Run("use env for minimal non-bootstrap", func(t *testing.T) {
-		kb, err := keyStore.OCR().Create()
-		require.NoError(t, err)
 		s := `
 		type               = "offchainreporting"
 		schemaVersion      = 1
@@ -521,11 +525,6 @@ ds1 -> ds1_parse;
 """
 `
 		s = fmt.Sprintf(s, cltest.NewEIP55Address(), "http://blah.com", "")
-		tAddress := ethkey.EIP55AddressFromAddress(transmitterAddress)
-		config.Overrides.P2PBootstrapPeers = []string{"/dns4/chain.link/tcp/1234/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju", "/dns4/chain.link/tcp/1235/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju"}
-		config.Overrides.P2PV2Bootstrappers = []commontypes.BootstrapperLocator{}
-		config.Overrides.OCRKeyBundleID = null.NewString(kb.ID(), true)
-		config.Overrides.OCRTransmitterAddress = &tAddress
 		jb, err := ocr.ValidatedOracleSpecToml(cc, s)
 		require.NoError(t, err)
 		err = toml.Unmarshal([]byte(s), &jb)
@@ -539,8 +538,6 @@ ds1 -> ds1_parse;
 		assert.Equal(t, models.Interval(20000000000), jb.OCROracleSpec.BlockchainTimeout)
 		assert.Equal(t, models.Interval(cltest.MustParseDuration(t, "1s")), jb.MaxTaskDuration)
 
-		// Required to create job spawner delegate.
-		config.Overrides.P2PListenPort = null.IntFrom(2000)
 		lggr := logger.TestLogger(t)
 		pw := ocrcommon.NewSingletonPeerWrapper(keyStore, config, db, lggr)
 		require.NoError(t, pw.Start(testutils.Context(t)))
@@ -575,8 +572,6 @@ ds1 -> ds1_parse;
 		require.NoError(t, err)
 		assert.Equal(t, jb.MaxTaskDuration, models.Interval(cltest.MustParseDuration(t, "1s")))
 
-		// Required to create job spawner delegate.
-		config.Overrides.P2PListenPort = null.IntFrom(2000)
 		lggr := logger.TestLogger(t)
 		pw := ocrcommon.NewSingletonPeerWrapper(keyStore, config, db, lggr)
 		require.NoError(t, pw.Start(testutils.Context(t)))
@@ -605,8 +600,6 @@ ds1 -> ds1_parse;
 		err = jobORM.CreateJob(&jb)
 		require.NoError(t, err)
 
-		// Required to create job spawner delegate.
-		config.Overrides.P2PListenPort = null.IntFrom(2000)
 		lggr := logger.TestLogger(t)
 		pw := ocrcommon.NewSingletonPeerWrapper(keyStore, config, db, lggr)
 		require.NoError(t, pw.Start(testutils.Context(t)))
@@ -637,9 +630,6 @@ ds1 -> ds1_parse;
 		err = jobORM.CreateJob(jb)
 		require.NoError(t, err)
 
-		// Required to create job spawner delegate.
-		config.Overrides.P2PListenPort = null.IntFrom(2000)
-		config.Overrides.P2PV2Bootstrappers = []commontypes.BootstrapperLocator{}
 		lggr := logger.TestLogger(t)
 		pw := ocrcommon.NewSingletonPeerWrapper(keyStore, config, db, lggr)
 		require.NoError(t, pw.Start(testutils.Context(t)))