+++ core/web/resolver/resolver_test.go
@@ -14,11 +14,11 @@ import (
 	evmConfigMocks "github.com/smartcontractkit/chainlink/core/chains/evm/config/mocks"
 	evmORMMocks "github.com/smartcontractkit/chainlink/core/chains/evm/mocks"
 	txmgrMocks "github.com/smartcontractkit/chainlink/core/chains/evm/txmgr/mocks"
-	configMocks "github.com/smartcontractkit/chainlink/core/config/mocks"
 	coremocks "github.com/smartcontractkit/chainlink/core/internal/mocks"
 	"github.com/smartcontractkit/chainlink/core/internal/testutils"
 	"github.com/smartcontractkit/chainlink/core/internal/testutils/evmtest"
 	"github.com/smartcontractkit/chainlink/core/logger/audit"
+	chainlinkMocks "github.com/smartcontractkit/chainlink/core/services/chainlink/mocks"
 	feedsMocks "github.com/smartcontractkit/chainlink/core/services/feeds/mocks"
 	jobORMMocks "github.com/smartcontractkit/chainlink/core/services/job/mocks"
 	keystoreMocks "github.com/smartcontractkit/chainlink/core/services/keystore/mocks"
@@ -38,7 +38,7 @@ type mocks struct {
 	sessionsORM *sessionsMocks.ORM
 	pipelineORM *pipelineMocks.ORM
 	feedsSvc    *feedsMocks.Service
-	cfg         *configMocks.GeneralConfig
+	cfg         *chainlinkMocks.GeneralConfig
 	scfg        *evmConfigMocks.ChainScopedConfig
 	ocr         *keystoreMocks.OCR
 	ocr2        *keystoreMocks.OCR2
@@ -96,7 +96,7 @@ func setupFramework(t *testing.T) *gqlTestFramework {
 		feedsSvc:    feedsMocks.NewService(t),
 		sessionsORM: sessionsMocks.NewORM(t),
 		pipelineORM: pipelineMocks.NewORM(t),
-		cfg:         configMocks.NewGeneralConfig(t),
+		cfg:         chainlinkMocks.NewGeneralConfig(t),
 		scfg:        evmConfigMocks.NewChainScopedConfig(t),
 		ocr:         keystoreMocks.NewOCR(t),
 		ocr2:        keystoreMocks.NewOCR2(t),