+++ docs/CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- unreleased -->
 ## [dev]
 
-...
+### Removed
+- Configuration with legacy environment variables is no longer supported. TOML is required.
 
 ## 1.13.0 - UNRELEASED
 
@@ -28,7 +29,7 @@ of any values provided via configuration files.
 
 ### Removed
 
-- Terra is no longer supported
+- Terra is no longer supported 
 
 <!-- unreleasedstop -->